Added fixes after review)

+++ src/main/java/core/basesyntax/Square.java
@@ -3,7 +3,7 @@ package core.basesyntax;
 public class Square extends Figure {
     private final double side;
 
-    public Square(String color, int side) {
+    public Square(String color, double side) {
         super(color);
         this.side = side;
     }
